refactor: 🛠  refactored and splited code for better readbility
